Clarified usage with less (and other transpilers)
The attribute selector injected by the scoped-style-loader interacts badly with some less operators (like the & parent selector), which can result in hard to debug webpack errors. This commit updates the readme to suggest marking the less-loader (and other similar loaders) as normal rather than post loaders to avoid this error. It also suggests a workaround for using the `>>>` operator in less.